+++ ubpe_cython/ubpe/interface.pxd
@@ -9,11 +9,11 @@ cdef extern from "ubpe_classic.hpp" namespace "ubpe":
     cdef cppclass UbpeClassic[DocType, TokenType]:
         UbpeClassic(uint32_t, uint32_t) except +
         UbpeClassic(uint32_t, uint32_t, map[TokenType, uint32_t]) except +
-        UbpeClassic(uint32_t, uint32_t, map[TokenType, uint32_t], map[uint32_t, TokenType], map[vector[uint32_t], uint32_t], map[uint32_t, vector[uint32_t]], map[uint32_t, float]) except +
+        UbpeClassic(uint32_t, uint32_t, map[TokenType, uint32_t], map[uint32_t, TokenType], map[vector[uint32_t], uint32_t], map[uint32_t, vector[uint32_t]], map[uint32_t, double]) except +
 
         void fit(const vector[DocType]& corpus, uint32_t n_candidates, bool rearrange_tokens) except +
 
-        vector[pair[vector[uint32_t], float]] encode(const DocType& doc, uint8_t) except +
+        vector[pair[vector[uint32_t], double]] encode(const DocType& doc, uint8_t) except +
 
         DocType decode(const vector[uint32_t]& tokens) except +
 
@@ -21,7 +21,7 @@ cdef extern from "ubpe_classic.hpp" namespace "ubpe":
 
         map[uint32_t, vector[uint32_t]] getBackwardMapper()
 
-        map[uint32_t, float] getTokensWeights()
+        map[uint32_t, double] getTokensWeights()
 
         map[TokenType, uint32_t] getAlphabet()
 
@@ -33,11 +33,11 @@ cdef extern from "ubpe.hpp" namespace "ubpe":
     cdef cppclass Ubpe[DocType, TokenType]:
         Ubpe(uint32_t, uint32_t) except +
         Ubpe(uint32_t, uint32_t, map[TokenType, uint32_t]) except +
-        Ubpe(uint32_t, uint32_t, map[TokenType, uint32_t], map[uint32_t, TokenType], map[vector[uint32_t], uint32_t], map[uint32_t, vector[uint32_t]], map[uint32_t, float]) except +
+        Ubpe(uint32_t, uint32_t, map[TokenType, uint32_t], map[uint32_t, TokenType], map[vector[uint32_t], uint32_t], map[uint32_t, vector[uint32_t]], map[uint32_t, double]) except +
 
         void fit(const vector[DocType]& corpus, uint32_t n_candidates, bool rearrange_tokens) except +
 
-        vector[pair[vector[uint32_t], float]] encode(const DocType& doc, uint8_t n_candidates) except +
+        vector[pair[vector[uint32_t], double]] encode(const DocType& doc, uint8_t n_candidates) except +
 
         DocType decode(const vector[uint32_t]& tokens) except +
 
@@ -45,7 +45,7 @@ cdef extern from "ubpe.hpp" namespace "ubpe":
 
         map[uint32_t, vector[uint32_t]] getBackwardMapper()
 
-        map[uint32_t, float] getTokensWeights()
+        map[uint32_t, double] getTokensWeights()
 
         map[TokenType, uint32_t] getAlphabet()
